pmgr: WIP: multi-die support for t6002

+++ src/pmgr.h
@@ -7,8 +7,8 @@
 
 int pmgr_init(void);
 
-int pmgr_power_enable(u16 id);
-int pmgr_power_disable(u16 id);
+int pmgr_power_enable(u32 id);
+int pmgr_power_disable(u32 id);
 
 int pmgr_adt_power_enable(const char *path);
 int pmgr_adt_power_disable(const char *path);
